+++ src/sbi/timer.rs
@@ -3,11 +3,11 @@ use crate::config::CLOCK_FREQ;
 use riscv::register::time;
 
 // trigger per 1ms
-pub const TICKS_PER_SECOND  : usize = 1000;
-pub const MILLI_PER_SECOND  : usize = 1000;
+pub const TICKS_PER_SECOND  : u64 = 1000;
+pub const MILLI_PER_SECOND  : u64 = 1000;
 
-pub fn get_time() -> usize {
-    time::read()
+pub fn get_time() -> u64 {
+    time::read() as u64
 }
 
 pub fn reset_timer_trigger() {
